ラジオボタン、申請の必要・不要の定数をutilsに移動
